Move location fetching into its own hook
